requirements.txt
asgiref==3.10.0
asttokens==2.4.1
attrs==25.4.0
autobahn==22.7.1
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
executing==2.2.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
iniconfig==2.3.0
model-bakery==1.20.5
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-django==4.11.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
zope.interface==8.0.1
reportlab==4.2.5
